# API-document.yaml
openapi: 3.0.0
servers:
  - description: 本地服务器
    url: 'http://localhost:8000'
  - description: 线上服务器
    url: 'https://www.heerkirov.com'
info:
  title: Basic Service REST API
  version: 0.2.0
  description: |
    提供基础的用户认证和管理服务。
    ### 使用说明
    API分为三部分。
    * 用户API：用户注册登录，查阅用户信息，查阅使用的App信息。
    * 管理API：具有管理权限的用户修改系统设置，管理注册码、用户、App。
    * 应用程序对接API：提供给使用此服务的应用程序，进行校验对接，提供比用户校验API更多的功能。
    ### 用户认证
    系统只提供Bearer Token认证。在发送请求时，编写请求头`Authorization`: `Bearer {token}`。token使用API`/api/token/`获得。
    用户认证成功时，会继续后续的业务请求。
    ### 公共错误信息
    下面是几种几乎所有API都会存在的错误返回信息，在这里统一列出。
    1. 401 Unauthorized
      * 没有获得Authorization请求头，请求头无法被正确解析，或请求头不是合法的Bearer格式
      * token不存在，已经过期，或已被销毁
    2. 403 Forbidden
      * 当前的权限等级不足
      * 不止需要登录，还需要更高的权限，比如管理员权限
    3. 400 Bad Request
      * Request Body的内容不是合法的json数据
      * Body的内容无法满足API描述的需要
      * 只有无法被前端验证的错误才会在文档中列出。能够在前端检查中指出的错误可能有错误代码，但不会列出。
    ### App接入指南
    Basic Service向注册在案的App提供token接入功能。这分两个部分。
    #### 配置App信息
    1. 调用[/api/admin/app/](#/Admin%20App/post_api_admin_app_)API创建一个新的App。
    2. 调用[/api/admin/app/{app-id}/secret/](#/Admin%20App/get_api_admin_app__app_id__secret_)API，取得此App的密钥。
    3. 将密钥配置在App后端系统中。
    #### 使用Token认证功能
    1. App的用户使用[/api/token/](#/Token/post_api_token_)API来申请token，而无需在App的后端编写token申请相关的API。
    2. 后端的其他业务端点按照普通的token认证端点来编写。
    3. 在每个需要验证用户身份的逻辑点，后端调用[/api/interface/verify/](#/Interface/post_api_interface_verify_)API来验证token是否正确。
...
paths:
  /api/token/:
    post:
      tags: [Token]
      summary: "生成新Token"
      requestBody:
        content:
          application/json:
            schema: {$ref: '#/components/schemas/TokenCreateBody'}
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TokenResBody'}
        '401':
          description: Unauthorized
          content:
            plain/text:
              example:
                'Password wrong': 密码错误
                'User not exist': 该用户不存在
                'User not enabled': 该用户已被停用
  /api/token/{token}/:
    get:
      tags: [Token]
      summary: "查询Token的状态"
      parameters:
        - name: token
          in: path
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TokenResBody'}
        '404':
          description: Not Found
    put:
      tags: [Token]
      summary: "对Token的持续时间进行刷新"
      parameters:
        - name: token
          in: path
          required: true
          schema: {type: string}
      requestBody:
        content:
          application/json:
            schema: {$ref: '#/components/schemas/TokenUpdateBody'}
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TokenResBody'}
        '404':
          description: Not Found
...
  /api/user/:
    get:
      tags: [User]
      summary: "查看用户信息"
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema: {$ref: '#/components/schemas/UserResBody'}
    post:
      tags: [User]
      summary: "更改用户信息"
      requestBody:
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UserUpdateBody'}
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema: {$ref: '#/components/schemas/UserResBody'}
    put:
      tags: [User]
      summary: "更改用户信息(同POST)"
      requestBody:
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UserUpdateBody'}
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema: {$ref: '#/components/schemas/UserResBody'}
  /api/user/password/:
    post:
      tags: [User]
      summary: "修改用户密码"
      requestBody:
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PasswordUpdateBody'}
      responses:
        '200':
          description: Ok
        '401':
          description: Unauthorized
          content:
            text/plain:
              example:
                'Password wrong': 旧密码错误
  /api/user/cover/:
    post:
      tags: [User]
      summary: "上传用户头像"
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: object
                  description: 要上传的文件流
      responses:
        '200':
          description: Ok
        '400':
          description: Bad Request - 文件转换出现问题，或文件类型不正确
...
  /api/admin/user/:
    get:
      tags: ['Admin User']
      summary: "获得用户列表"
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/AdminUserResBody'}
    post:
      tags: ['Admin User']
      summary: "创建新用户"
      requestBody:
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AdminUserCreateBody'}
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminUserResBody'}
  /api/admin/user/{username}/:
    get:
      tags: ['Admin User']
      summary: "获得用户详情"
      parameters:
        - name: username
          required: true
          in: path
          schema: {type: string}
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminUserResBody'}
        '404':
          description: Not Found
    put:
      tags: ['Admin User']
      summary: "变更用户状态"
      parameters:
        - name: username
          required: true
          in: path
          schema: {type: string}
      requestBody:
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AdminUserUpdateBody'}
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminUserResBody'}
        '404':
          description: Not Found
    delete:
      tags: ['Admin User']
      summary: "删除用户"
      parameters:
        - name: username
          required: true
          in: path
          schema: {type: string}
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
  /api/admin/user/{username}/password/:
    put:
      tags: ['Admin User']
      summary: "修改用户密码"
      parameters:
        - name: username
          required: true
          in: path
          schema: {type: string}
      requestBody:
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AdminPasswordUpdateBody'}
      responses:
        '200':
          description: Ok
        '404':
          description: Not Found
...
components:
  schemas:
    TokenCreateBody:
      type: object
      required: [username, password]
      description: |
        token认证需要用户名和密码。除此之外，还可以设置token的有效时长。
        * 如果不设置任何与时长有关的参数，那么持续时长将是系统的默认时长；
        * 如果将`effective_unlimit`设置为`true`，那么将使用系统规定的能使用的最长时长，如果最长时长是无限，那么可以设置无限长的token；
        * 如果设置`effective`参数，那么将请求这个长度的持续时长，但如果这个时长超过了最大时长，还是会被设定为最大时长。
      properties:
        username:
          type: string
          description: 用户名。
        password:
          type: string
          description: 密码
        effective:
          type: number
          description: 申请的token有效时长，单位毫秒。
        effective_unlimit:
          type: boolean
          description: 请求无限长度的token时长。
          default: false
    TokenUpdateBody:
      type: object
      properties:
        effective:
          type: number
          description: token的有效时长
    TokenResBody:
      type: object
      properties:
        key:
          type: string
          description: token
        username:
          type: string
          description: 用户名
        expire_time:
          type: string
          format: datetime
          description: token过期的时间。`null`表示不会过期。
        create_time:
          type: string
          format: datetime
          description: token创建的时间
        update_time:
          type: string
          format: datetime
          description: token上次更新的时间
...
    UserUpdateBody:
      type: object
      required: [name]
      properties:
        name:
          type: string
          description: 用户显示名
    UserResBody:
      type: object
      properties:
        username:
          type: string
          description: 用户名
        name:
          type: string
          description: 显示名
        cover:
          type: string
          description: 用户头像文件的文件名，可能为`null`。
        is_staff:
          type: boolean
          description: 是否为系统管理员
        last_login:
          type: string
          format: datetime
          description: 上次登录时间。如果为`null`表示用户还没有登录过。
        last_login_ip:
          type: string
          description: 上次登录的IP。如果为`null`表示用户还没有登录过，或者识别不了IP。
        create_time:
          type: string
          format: datetime
          description: 用户创建的时间
        create_path:
          type: string
          enum: [System, Admin, Code, Public]
          description: |
            用户创建的途径。
            * System: 用户由系统创建。只有初始化系统时创建的初始用户会通过此途径创建。
            * Admin: 用户由系统管理员在后台创建。
            * Code: 用户使用注册码自行注册。
            * Public: 用户通过开放注册自行注册。
    PasswordUpdateBody:
      type: object
      required: [old_password, new_password]
      properties:
        old_password:
          type: string
          description: 旧密码
        new_password:
          type: string
          description: 新密码
...
    AdminUserCreateBody:
      type: object
      required: [username, password, name, is_staff]
      properties:
        username:
          type: string
          description: 用户名，不允许重复，符合`[A-Za-z0-9_-]+`字符集。
        password:
          type: string
          description: 密码
        name:
          type: string
          description: 用户显示的名称
        is_staff:
          type: boolean
          description: 是否是管理员用户
    AdminUserUpdateBody:
      type: object
      required: [enable]
      properties:
        enable:
          type: boolean
          description: 该用户是否可用
    AdminUserResBody:
      allOf:
        - $ref: '#/components/schemas/UserResBody'
        - type: object
          properties:
            enable:
              type: boolean
              description: 用户是否可用
    AdminPasswordUpdateBody:
      type: object
      required: [new_password]
      properties:
        new_password:
          type: string
          description: 新密码
...